Refactors manage.py
