+++ stream_alert_cli/kinesis/handler.py
@@ -16,38 +16,74 @@ limitations under the License.
 from stream_alert.shared.logger import get_logger
 from stream_alert_cli.helpers import tf_runner
 from stream_alert_cli.terraform.generate import terraform_generate_handler
+from stream_alert_cli.utils import add_clusters_arg, set_parser_epilog, CliCommand
 
 LOGGER = get_logger(__name__)
 
 
-def kinesis_handler(options, config):
-    """Main handler for the Kinesis parser
+class KinesisCommand(CliCommand):
 
-    Args:
-        options (argparse.Namespace): Parsed arguments
-        config (CLIConfig): Loaded StreamAlert config
+    @classmethod
+    def setup_subparser(cls, subparser):
+        """Add kinesis subparser: manage.py kinesis [options]"""
+        set_parser_epilog(
+            subparser,
+            epilog=(
+                '''\
+                Example:
 
-    Returns:
-        bool: False if errors occurred, True otherwise
-    """
-    enable = options.action == 'enable-events'
-    LOGGER.info('%s Kinesis Events', 'Enabling' if enable else 'Disabling')
+                    manage.py kinesis disable-events --clusters corp prod
+                '''
+            )
+        )
 
-    for cluster in options.clusters or config.clusters():
-        if 'kinesis_events' in config['clusters'][cluster]['modules']:
-            config['clusters'][cluster]['modules']['kinesis_events']['enabled'] = enable
+        actions = ['disable-events', 'enable-events']
+        subparser.add_argument(
+            'action',
+            metavar='ACTION',
+            choices=actions,
+            help='One of the following actions to be performed: {}'.format(', '.join(actions))
+        )
 
-    config.write()
+        # Add the option to specify cluster(s)
+        add_clusters_arg(subparser)
 
-    if options.skip_terraform:
-        return True  # not an error
+        subparser.add_argument(
+            '-s',
+            '--skip-terraform',
+            action='store_true',
+            help='Only update the config options and do not run Terraform'
+        )
 
-    if not terraform_generate_handler(config):
-        return False
+    @classmethod
+    def handler(cls, options, config):
+        """Main handler for the Kinesis parser
 
-    return tf_runner(
-        action='apply',
-        targets=[
-            'module.{}_{}'.format('kinesis_events', cluster) for cluster in config.clusters()
-        ]
-    )
+        Args:
+            options (argparse.Namespace): Parsed arguments
+            config (CLIConfig): Loaded StreamAlert config
+
+        Returns:
+            bool: False if errors occurred, True otherwise
+        """
+        enable = options.action == 'enable-events'
+        LOGGER.info('%s Kinesis Events', 'Enabling' if enable else 'Disabling')
+
+        for cluster in options.clusters or config.clusters():
+            if 'kinesis_events' in config['clusters'][cluster]['modules']:
+                config['clusters'][cluster]['modules']['kinesis_events']['enabled'] = enable
+
+        config.write()
+
+        if options.skip_terraform:
+            return True  # not an error
+
+        if not terraform_generate_handler(config):
+            return False
+
+        return tf_runner(
+            action='apply',
+            targets=[
+                'module.{}_{}'.format('kinesis_events', cluster) for cluster in config.clusters()
+            ]
+        )
