+++ stream_alert_cli/test/handler.py
@@ -35,34 +35,172 @@ from stream_alert.shared import rule
 from stream_alert.shared.logger import get_logger
 from stream_alert.shared.stats import get_rule_stats
 from stream_alert_cli.helpers import check_credentials
+from stream_alert_cli.test import DEFAULT_TEST_FILES_DIRECTORY
 from stream_alert_cli.test.format import format_green, format_red, format_underline, format_yellow
 from stream_alert_cli.test.mocks import mock_lookup_table_results, mock_threat_intel_query_results
 from stream_alert_cli.test.results import TestEventFile, TestResult
+from stream_alert_cli.utils import UniqueSetAction, generate_subparser
 
 LOGGER = get_logger(__name__)
 
 
-def test_handler(options, config):
-    """Handler for starting the test framework
+class TestCommand(object):
 
-    Args:
-        options (argparse.Namespace): Parsed arguments
-        config (CLIConfig): Loaded StreamAlert config
+    @classmethod
+    def setup_subparser(cls, subparser):
+        """Add the test subparser: manage.py test"""
+        test_subparsers = subparser.add_subparsers()
 
-    Returns:
-        bool: False if errors occurred, True otherwise
-    """
-    result = True
-    opts = vars(options)
-    repeat = opts.get('repeat', 1)
-    for i in range(repeat):
-        if repeat != 1:
-            print('\nRepetition #', i+1)
-        result = result and TestRunner(options, config).run()
+        cls._setup_test_classifier_subparser(test_subparsers)
+        cls._setup_test_rules_subparser(test_subparsers)
+        cls._setup_test_live_subparser(test_subparsers)
 
-    if opts.get('stats'):
-        print(get_rule_stats())
-    return result
+    @classmethod
+    def _setup_test_classifier_subparser(cls, subparsers):
+        """Add the test validation subparser: manage.py test classifier [options]"""
+        test_validate_parser = generate_subparser(
+            subparsers,
+            'classifier',
+            description='Validate defined log schemas using integration test files',
+            subcommand=True
+        )
+
+        cls._add_default_test_args(test_validate_parser)
+
+    @classmethod
+    def _setup_test_rules_subparser(cls, subparsers):
+        """Add the test rules subparser: manage.py test rules [options]"""
+        test_rules_parser = generate_subparser(
+            subparsers,
+            'rules',
+            description='Test rules using integration test files',
+            subcommand=True
+        )
+
+        # Flag to run additional stats during testing
+        test_rules_parser.add_argument(
+            '-s',
+            '--stats',
+            action='store_true',
+            help='Enable outputing of statistical information on rules that run'
+        )
+
+        # flag to run these tests a given number of times
+        test_rules_parser.add_argument(
+            '-n',
+            '--repeat',
+            default=1,
+            type=cls._create_repetition_validator(test_rules_parser),
+            help='Number of times to repeat the tests, to be used as a form performance testing'
+        )
+
+        cls._add_default_test_args(test_rules_parser)
+
+    @staticmethod
+    def _create_repetition_validator(test_rules_parser):
+        def _validate_repitition(val):
+            """Make sure the input is between 1 and 1000"""
+            err = ('Invalid repitition value [{}]. Must be an integer between 1 '
+                   'and 1000').format(val)
+            try:
+                count = int(val)
+            except TypeError:
+                raise test_rules_parser.error(err)
+
+            if not 1 <= count <= 1000:
+                raise test_rules_parser.error(err)
+
+            return count
+
+        return _validate_repitition
+
+    @classmethod
+    def _setup_test_live_subparser(cls, subparsers):
+        """Add the test live subparser: manage.py test live [options]"""
+        test_live_parser = generate_subparser(
+            subparsers,
+            'live',
+            description='Run end-to-end tests that will attempt to send alerts to each rule\'s outputs',
+            subcommand=True
+        )
+
+        cls._add_default_test_args(test_live_parser)
+
+    @staticmethod
+    def _add_default_test_args(test_parser):
+        """Add the default arguments to the test parsers"""
+        test_filter_group = test_parser.add_mutually_exclusive_group(required=False)
+
+        # add the optional ability to test against a rule/set of rules
+        test_filter_group.add_argument(
+            '-f',
+            '--test-files',
+            dest='files',
+            metavar='FILENAMES',
+            nargs='+',
+            help='One or more file to test, separated by spaces',
+            action=UniqueSetAction,
+            default=set()
+        )
+
+        # add the optional ability to test against a rule/set of rules
+        test_filter_group.add_argument(
+            '-r',
+            '--test-rules',
+            dest='rules',
+            nargs='+',
+            help='One or more rule to test, separated by spaces',
+            action=UniqueSetAction,
+            default=set()
+        )
+
+        # add the optional ability to change the test files directory
+        test_parser.add_argument(
+            '-d',
+            '--files-dir',
+            help='Path to directory containing test files',
+            default=DEFAULT_TEST_FILES_DIRECTORY
+        )
+
+        # Add the optional ability to log verbosely or use quite logging for tests
+        verbose_group = test_parser.add_mutually_exclusive_group(required=False)
+
+        verbose_group.add_argument(
+            '-v',
+            '--verbose',
+            action='store_true',
+            help='Output additional information during testing'
+        )
+
+        verbose_group.add_argument(
+            '-q',
+            '--quiet',
+            action='store_true',
+            help='Suppress output for passing tests, only logging if there is a failure'
+        )
+
+    @classmethod
+    def handler(cls, options, config):
+        """Handler for starting the test framework
+
+        Args:
+            options (argparse.Namespace): Parsed arguments
+            config (CLIConfig): Loaded StreamAlert config
+
+        Returns:
+            bool: False if errors occurred, True otherwise
+        """
+        result = True
+        opts = vars(options)
+        repeat = opts.get('repeat', 1)
+        for i in range(repeat):
+            if repeat != 1:
+                print('\nRepetition #', i+1)
+            result = result and TestRunner(options, config).run()
+
+        if opts.get('stats'):
+            print(get_rule_stats())
+        return result
 
 
 class TestRunner(object):
@@ -111,11 +249,10 @@ class TestRunner(object):
     def _run_rules_engine(self, record):
         """Create a fresh rules engine and process the record, returning the result"""
         with patch.object(rules_engine.ThreatIntel, '_query') as ti_mock, \
-             patch.object(rules_engine.LookupTables, 'get_instance') as lt_mock, \
-             patch.object(rules_engine, 'AlertForwarder'), \
-             patch.object(rules_engine, 'RuleTable') as rule_table, \
-             patch('rules.helpers.base.random_bool', return_value=True):
-
+                patch.object(rules_engine.LookupTables, 'get_instance') as lt_mock, \
+                patch.object(rules_engine, 'AlertForwarder'), \
+                patch.object(rules_engine, 'RuleTable') as rule_table, \
+                patch('rules.helpers.base.random_bool', return_value=True):
             # Emptying out the rule table will force all rules to be unstaged, which causes
             # non-required outputs to get properly populated on the Alerts that are generated
             # when running the Rules Engine.
@@ -302,7 +439,8 @@ class TestRunner(object):
 
         # Log any errors for filtered items that do not exist
         for basename in files_filter:
-            self._append_error('No test event file found with base name \'{}\''.format(basename))
+            self._append_error(
+                'No test event file found with base name \'{}\''.format(basename))
 
     def _setup_s3_mock(self, data):
         self._s3_mocker.return_value.Bucket.return_value.download_fileobj = (
@@ -521,7 +659,8 @@ class TestRunner(object):
 
         optional_keys = {'compress', 'trigger_rules', 'validate_schema_only'}
 
-        key_diff = test_event_keys.difference(required_keys | optional_keys | acceptable_data_keys)
+        key_diff = test_event_keys.difference(
+            required_keys | optional_keys | acceptable_data_keys)
 
         # Log a warning if there are extra keys declared in the test log
         if key_diff:
